Catching everything up

+++ ExpenseManager.Business/Services/ExpenseDataServiceExtensions.cs
@@ -8,13 +8,13 @@ namespace ExpenseManager.Business.Services
 {
     public static class ExpenseDataServiceExtensions
     {
-        public static IEnumerable<Expense> GetExpensesAboveValueForYear(this ExpenseDataService expenseDataService, double value, DateTime year)
+        public static IEnumerable<Expense> GetExpensesAboveValueForYear(this ExpenseDataService expenseDataService, decimal value, DateTime year)
         {
             return expenseDataService.GetExpensesByYear(year).Where(x => x.Value > value);
         }
 
         public static IEnumerable<Expense> GetExpensesAboveValueForMonthAndYear(
-            this ExpenseDataService expenseDataService, double value, DateTime monthAndYear)
+            this ExpenseDataService expenseDataService, decimal value, DateTime monthAndYear)
         {
             return expenseDataService.GetExpensesByMonthAndYear(monthAndYear).Where(x => x.Value > value);
         }
